todolist v0.3 fix types imports
