+++ src/types.ts
@@ -1,0 +1,25 @@
+export interface TodoRequest {
+  title?: string;
+  isDone?: boolean; // изменение статуса задачи происходит через этот флаг
+}
+
+export interface TodoInfo {
+  all: number;
+  completed: number;
+  inWork: number;
+}
+
+export interface Todo {
+  id: number;
+  title: string;
+  created: string; // ISO date string
+  isDone: boolean;
+}
+
+export interface MetaResponse<T, N> {
+  data: T[];
+  info?: N;
+  meta: {
+    totalAmount: number;
+  };
+}
